[UHYU-145] feat: brand-grid 공통 컴포넌트 화 및 배럴 익스포트 패턴 적용
