+++ src/pages/user/extra-info/types/index.ts
@@ -1,14 +1,3 @@
-export interface Brand {
-  id: string;
-  name: string;
-  color: string;
-  bgColor: string;
-  textColor: string;
-  logo?: string;
-  logoAlt?: string;
-  imagePath: string;
-}
-
 export interface MembershipGrade {
   value: string;
   label: string;
@@ -62,21 +51,6 @@ export interface CurrentStepProps {
 
 export interface CompletedStepsProps {
   completedSteps: CompletedStep[];
-}
-
-export interface BrandLogoProps {
-  brand: Brand;
-  isSelected: boolean;
-  onClick?: () => void;
-  delay?: number;
-  disabled?: boolean;
-}
-
-export interface BrandGridProps {
-  selectedBrands: string[];
-  onBrandToggle?: (brandId: string) => void;
-  title: string;
-  disabled?: boolean;
 }
 
 export interface ActionButtonsProps {
