+++ src/shared/components/brand_grid/BrandGrid.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { BrandLogo } from './BrandLogo';
-import { BRANDS } from '../constants';
-import { type BrandGridProps } from '../types';
+import { BrandLogo } from '@shared/components/brand_grid/BrandLogo';
+import { BRANDS } from '@shared/components/brand_grid/constants';
+import { type BrandGridProps } from '@shared/components/brand_grid/brand.type';
 
 export const BrandGrid: React.FC<BrandGridProps> = ({
   selectedBrands,
