+++ src/shared/components/brand_grid/brand.type.ts
@@ -1,0 +1,25 @@
+export interface Brand {
+  id: string;
+  name: string;
+  color: string;
+  bgColor: string;
+  textColor: string;
+  logo?: string;
+  logoAlt?: string;
+  imagePath: string;
+}
+
+export interface BrandGridProps {
+  selectedBrands: string[];
+  onBrandToggle?: (brandId: string) => void;
+  title: string;
+  disabled?: boolean;
+}
+
+export interface BrandLogoProps {
+  brand: Brand;
+  isSelected: boolean;
+  onClick?: () => void;
+  delay?: number;
+  disabled?: boolean;
+}
