+++ src/shared/components/brand_grid/index.tsx
@@ -1,0 +1,8 @@
+export { BrandGrid } from './BrandGrid';
+export { BrandLogo } from './BrandLogo';
+
+export type { Brand, BrandGridProps, BrandLogoProps } from './brand.type';
+
+export { BRANDS } from './constants';
+
+export { BrandGrid as default } from './BrandGrid';
